rename package sign to signature
